+++ src/main/java/com/mentalfrostbyte/jello/module/impl/movement/spider/VanillaSpider.java
@@ -2,7 +2,7 @@ package com.mentalfrostbyte.jello.module.impl.movement.spider;
 
 import com.mentalfrostbyte.jello.event.impl.player.movement.EventMove;
 import com.mentalfrostbyte.jello.module.Module;
-import com.mentalfrostbyte.jello.module.ModuleCategory;
+import com.mentalfrostbyte.jello.module.data.ModuleCategory;
 import com.mentalfrostbyte.jello.module.settings.impl.NumberSetting;
 import team.sdhq.eventBus.annotations.EventTarget;
 
@@ -16,7 +16,7 @@ public class VanillaSpider extends Module {
     public void EventMove(EventMove event) {
         if (this.isEnabled()) {
             if (mc.player.collidedHorizontally) {
-                event.setY((double)this.getNumberValueBySettingName("Motion"));
+                event.setY(this.getNumberValueBySettingName("Motion"));
             }
         }
     }
